restructure: organised files properly
