+++ packages/shopstr-mcp/src/tools/get-reviews.ts
@@ -19,8 +19,8 @@ import {
   createRelayUnavailableResponse,
   createValidationErrorResponse,
   getDataFreshness,
-} from "./common.js";
-import type { CoreToolContext } from "./context.js";
+} from "./utils/common.js";
+import type { CoreToolContext } from "./utils/context.js";
 
 export const getReviewsInputSchema = {
   productId: z
