+++ packages/shopstr-mcp/src/tools/search-products.ts
@@ -14,8 +14,8 @@ import {
   createRelayUnavailableResponse,
   createValidationErrorResponse,
   getDataFreshness,
-} from "./common.js";
-import type { CoreToolContext } from "./context.js";
+} from "./utils/common.js";
+import type { CoreToolContext } from "./utils/context.js";
 
 export const searchProductsInputSchema = {
   keyword: z
