+++ packages/shopstr-mcp/test/tools/get-product-details.node.mjs
