+++ packages/shopstr-mcp/test/tools/get-reviews.node.mjs
