+++ packages/shopstr-mcp/test/tools/search-products.node.mjs
